[Refactor][l] Splitting main lib files into api and hybrid approach (db + api) files + refactroing DAGs with new library file locations
